crc checks are still failing

+++ includes/interfaces/data_package.hpp
@@ -5,7 +5,7 @@
 template <std::size_t size> struct data_package
 {
   std::array<std::uint8_t, size> voltage_bus{ 0 };
-  std::uint16_t hash{ 0 };
+  std::uint32_t hash{ 0 };
 };
 
 #endif // !DATA_PACKAGE_HPP
